requirements.txt
questionary>=2.0
rich>=13.0
jinja2>=3.1
pyyaml>=6.0
